% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_homography_algorithm.R
\name{ov_detect_court_plot}
\alias{ov_detect_court_plot}
\title{Plot the detected court homographies}
\usage{
ov_detect_court_plot(
  obj,
  index = 1,
  model.filter = NULL,
  plot.segments = FALSE,
  plot.all.segments = FALSE,
  plot.endpoints = FALSE,
  plot.all.endpoints = FALSE
)
}
\arguments{
\item{obj}{: as returned by \code{\link[=ov_detect_court]{ov_detect_court()}}}

\item{index}{numeric: which homographies to plot}

\item{model.filter}{: currently ignored}

\item{plot.segments, plot.all.segments, plot.endpoints, plot.all.endpoints}{logical: control aspects of plot appearance}
}
\description{
Plot the detected court homographies
}
