% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_homography_algorithm.R
\name{ov_plot_patterns}
\alias{ov_plot_patterns}
\title{Plot the detected court patterns}
\usage{
ov_plot_patterns(obj, index = 1)
}
\arguments{
\item{obj}{: as returned by \code{\link[=ov_detect_court]{ov_detect_court()}}}

\item{index}{numeric: which patterns to plot}
}
\description{
Plot the detected court patterns
}
