% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovcourt-package.R
\docType{package}
\name{ovcourt}
\alias{ovcourt-package}
\alias{ovcourt}
\title{\pkg{ovcourt}}
\description{
Court detection functions for images.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openvolley/ovcourt}
  \item Report bugs at \url{https://github.com/openvolley/ovcourt/issues}
}

}
\author{
\strong{Maintainer}: Adrien Ickowicz \email{adrien@untan.gl}

Authors:
\itemize{
  \item Ben Raymond
}

Other contributors:
\itemize{
  \item openvolley.org [originator]
}

}
